requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn==0.27.0

# Data Handling
pydantic==2.6.0
python-multipart==0.0.9
pyyaml==6.0.1

# Templates and Static Files
jinja2==3.1.3
aiofiles==23.2.1

# LLM Integration
zhipuai==2.0.1

# Development Tools
pytest==7.4.3
pytest-asyncio==0.23.2
